+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# [3.1.7] - 2021-03-11
+### Changed
+- Refactor many parts of daemon
+- Switch out session monitoring to logind-zbus 
+
 # [3.1.6] - 2021-03-11
 ### Changed
 - Graphics switching will now wait until all users logged out before switching